Moved DependenciesHasherTask
