+++ buildSrc/src/main/java/co/elastic/apm/compile/tools/plugins/subprojects/AarNoticeProviderPlugin.java
@@ -18,7 +18,7 @@ import co.elastic.apm.compile.tools.tasks.CreateDependenciesListTask;
 import co.elastic.apm.compile.tools.tasks.CreateNoticeTask;
 import co.elastic.apm.compile.tools.tasks.NoticeMergerTask;
 import co.elastic.apm.compile.tools.tasks.subprojects.CopySingleFileTask;
-import co.elastic.apm.compile.tools.tasks.subprojects.DependenciesHasherTask;
+import co.elastic.apm.compile.tools.tasks.dependencies.DependenciesHasherTask;
 import co.elastic.apm.compile.tools.tasks.subprojects.NoticeFilesCollectorTask;
 import co.elastic.apm.compile.tools.tasks.subprojects.PomLicensesCollectorTask;
 import co.elastic.apm.compile.tools.utils.Constants;
