+++ buildSrc/src/main/java/co/elastic/apm/compile/tools/plugins/subprojects/JarNoticeProviderPlugin.java
@@ -9,7 +9,7 @@ import java.util.List;
 import co.elastic.apm.compile.tools.tasks.CreateDependenciesListTask;
 import co.elastic.apm.compile.tools.tasks.CreateNoticeTask;
 import co.elastic.apm.compile.tools.tasks.NoticeMergerTask;
-import co.elastic.apm.compile.tools.tasks.subprojects.DependenciesHasherTask;
+import co.elastic.apm.compile.tools.tasks.dependencies.DependenciesHasherTask;
 import co.elastic.apm.compile.tools.tasks.subprojects.NoticeFilesCollectorTask;
 import co.elastic.apm.compile.tools.tasks.subprojects.PomLicensesCollectorTask;
